+++ business-core/src/main/java/com/tiktok/appevents/TTRequestBuilder.java
@@ -6,7 +6,6 @@
 
 package com.tiktok.appevents;
 
-import android.content.Context;
 import android.os.Build;
 import android.os.SystemClock;
 import android.text.TextUtils;
@@ -22,7 +21,6 @@ import com.tiktok.util.TimeUtil;
 import org.json.JSONObject;
 
 import java.util.Date;
-import java.util.Locale;
 
 class TTRequestBuilder {
     private static final String TAG = "TTRequestBuilder";
@@ -164,77 +162,9 @@ class TTRequestBuilder {
         }
     }
 
-    private static Locale getCurrentLocale() {
-        try {
-            Context context = TikTokBusinessSdk.getApplicationContext();
-            if (context != null) {
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-                    return context.getResources().getConfiguration().getLocales().get(0);
-                } else {
-                    // noinspection deprecation
-                    return context.getResources().getConfiguration().locale;
-                }
-            }
-        } catch (Throwable ignore) {
-        }
-        return null;
-    }
-
-    static String getBcp47Language() {
-        Locale loc = getCurrentLocale();
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-            return loc.toLanguageTag();
-        }
-
-        // we will use a dash as per BCP 47
-        final char SEP = '-';
-        String language = loc.getLanguage();
-        String region = loc.getCountry();
-        String variant = loc.getVariant();
-
-        // special case for Norwegian Nynorsk since "NY" cannot be a variant as per BCP 47
-        // this goes before the string matching since "NY" wont pass the variant checks
-        if (language.equals("no") && region.equals("NO") && variant.equals("NY")) {
-            language = "nn";
-            region = "NO";
-            variant = "";
-        }
-
-        if (language.isEmpty() || !language.matches("\\p{Alpha}{2,8}")) {
-            language = "und";       // Follow the Locale#toLanguageTag() implementation
-            // which says to return "und" for Undetermined
-        } else if (language.equals("iw")) {
-            language = "he";        // correct deprecated "Hebrew"
-        } else if (language.equals("in")) {
-            language = "id";        // correct deprecated "Indonesian"
-        } else if (language.equals("ji")) {
-            language = "yi";        // correct deprecated "Yiddish"
-        }
-
-        // ensure valid country code, if not well formed, it's omitted
-        if (!region.matches("\\p{Alpha}{2}|\\p{Digit}{3}")) {
-            region = "";
-        }
-
-        // variant subtags that begin with a letter must be at least 5 characters long
-        if (!variant.matches("\\p{Alnum}{5,8}|\\p{Digit}\\p{Alnum}{3}")) {
-            variant = "";
-        }
-
-        StringBuilder bcp47Tag = new StringBuilder(language);
-        if (!region.isEmpty()) {
-            bcp47Tag.append(SEP).append(region);
-        }
-        if (!variant.isEmpty()) {
-            bcp47Tag.append(SEP).append(variant);
-        }
-
-        return bcp47Tag.toString();
-    }
-
     private static JSONObject contextBuilderWithLocalAndLibrary(@Nullable TTIdentifierFactory.AdIdInfo adIdInfo) {
         JSONObject jsonObject = contextBuilder(adIdInfo, false);
-        JSON.putObject(jsonObject, "locale", getBcp47Language());
+        JSON.putObject(jsonObject, "locale", SystemInfoUtil.getLocale());
 
         JSONObject library = JSON.build();
         JSON.putObject(library, "name", "tiktok/" + SystemInfoUtil.getLibraryName());
@@ -306,7 +236,7 @@ class TTRequestBuilder {
             JSON.putObject(device, "ip", SystemInfoUtil.getLocalIpAddress());
             JSON.putObject(device, "network", SystemInfoUtil.getNetworkClass(TikTokBusinessSdk.getApplicationContext()));
             JSON.putObject(device, "session", TikTokBusinessSdk.getSessionID());
-            JSON.putObject(device, "locale", getBcp47Language());
+            JSON.putObject(device, "locale", SystemInfoUtil.getBcp47Language());
             JSON.putLong(device, "ts", System.currentTimeMillis() - SystemClock.elapsedRealtime());
             addDeviceInfo(device);
         } catch (Throwable ignore) {
@@ -316,7 +246,7 @@ class TTRequestBuilder {
 
     private static void addDeviceInfo(JSONObject device) {
         try {
-            JSON.putObject(device, "locale", getBcp47Language());
+            JSON.putObject(device, "locale", SystemInfoUtil.getBcp47Language());
             JSON.putInt(device, "screen_width", SystemInfoUtil.getsScreenWidth());
             JSON.putInt(device, "screen_height", SystemInfoUtil.getsScreenHeight());
             JSON.putDouble(device, "scale", SystemInfoUtil.getsDensity());
